% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RDD.R
\docType{methods}
\name{cache}
\alias{cache}
\alias{cache,RDD-method}
\title{Persist an RDD}
\usage{
cache(rdd)

\S4method{cache}{RDD}(rdd)
}
\arguments{
\item{rdd}{The RDD to cache}
}
\description{
Persist this RDD with the default storage level (MEMORY_ONLY).
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10, 2L)
cache(rdd)
}
}

