% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{cacheTable}
\alias{cacheTable}
\title{Cache Table}
\usage{
cacheTable(sqlctx, tableName)
}
\arguments{
\item{sqlctx}{SQLContext to use}

\item{tableName}{The name of the table being cached}
}
\value{
DataFrame
}
\description{
Caches the specified table in-memory.
}

