% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RDD.R
\docType{methods}
\name{count}
\alias{count}
\alias{count,RDD-method}
\alias{length,RDD-method}
\title{Return the number of elements in the RDD.}
\usage{
count(x)

\S4method{count}{RDD}(x)

\S4method{length}{RDD}(x)
}
\arguments{
\item{x}{The RDD to count}
}
\value{
number of elements in the RDD.
}
\description{
Return the number of elements in the RDD.

Return the number of elements in the RDD
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10)
count(rdd) # 10
length(rdd) # Same as count
}
}

