% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{jsonFile}
\alias{jsonFile}
\title{Create a DataFrame from a JSON file.}
\usage{
jsonFile(sqlctx, path)
}
\arguments{
\item{sqlctx}{SQLContext to use}

\item{path}{Path of file to read. A vector of multiple paths is allowed.}
}
\value{
DataFrame
}
\description{
Loads a JSON file (one object per line), returning the result as a DataFrame
It goes through the entire dataset once to determine the schema.
}

