% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RDD.R
\docType{methods}
\name{maximum}
\alias{maximum}
\alias{maximum,RDD}
\alias{maximum,RDD-method}
\title{Get the maximum element of an RDD.}
\usage{
maximum(rdd)

\S4method{maximum}{RDD}(rdd)
}
\arguments{
\item{rdd}{The RDD to get the maximum element from}
}
\description{
Get the maximum element of an RDD.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10)
maximum(rdd) # 10
}
}

