% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RDD.R
\docType{methods}
\name{minimum}
\alias{minimum}
\alias{minimum,RDD}
\alias{minimum,RDD-method}
\title{Get the minimum element of an RDD.}
\usage{
minimum(rdd)

\S4method{minimum}{RDD}(rdd)
}
\arguments{
\item{rdd}{The RDD to get the minimum element from}
}
\description{
Get the minimum element of an RDD.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10)
minimum(rdd) # 1
}
}

