% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{parquetFile}
\alias{parquetFile}
\title{Create a DataFrame from a Parquet file.}
\usage{
parquetFile(sqlctx, path)
}
\arguments{
\item{sqlctx}{SQLContext to use}

\item{path}{Path of file to read. A vector of multiple paths is allowed.}
}
\value{
DataFrame
}
\description{
Loads a Parquet file, returning the result as a DataFrame.
}

