% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RDD.R
\docType{methods}
\name{pipeRDD}
\alias{pipeRDD}
\alias{pipeRDD,RDD,character-method}
\title{Pipes elements to a forked externel process.}
\usage{
pipeRDD(rdd, command, ...)

\S4method{pipeRDD}{RDD,character}(rdd, command, env = character())
}
\arguments{
\item{rdd}{the RDD whose elements are piped to the forked externel process.}

\item{command}{the command to fork an externel process.}

\item{...}{other arguments passed to pipeRDD.}

\item{env}{character vector of name=value strings to set environment variables.}
}
\value{
a new RDD created by piping all elements to a forked externel process.
}
\description{
The same as `pipe()' in Spark.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10)
collect(pipeRDD(rdd, "cat")
Output: c("1", "2", ..., "10")
}
}

