% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RDD.R
\docType{methods}
\name{saveAsTextFile}
\alias{saveAsTextFile}
\alias{saveAsTextFile,RDD}
\alias{saveAsTextFile,RDD,character-method}
\title{Save this RDD as a text file, using string representations of elements.}
\usage{
saveAsTextFile(rdd, path)

\S4method{saveAsTextFile}{RDD,character}(rdd, path)
}
\arguments{
\item{rdd}{The RDD to save}

\item{path}{The directory where the splits of the text file are saved}
}
\description{
Save this RDD as a text file, using string representations of elements.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:3)
saveAsTextFile(rdd, "/tmp/sparkR-tmp")
}
}

