% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RDD.R
\docType{methods}
\name{sortByKey}
\alias{sortByKey}
\alias{sortByKey,RDD,RDD-method}
\alias{sortByKey,RDD-method}
\title{Sort a (k, v) pair RDD by k.}
\usage{
sortByKey(rdd, ascending = TRUE, numPartitions = 1L)

\S4method{sortByKey}{RDD}(rdd, ascending = TRUE,
  numPartitions = SparkR::numPartitions(rdd))
}
\arguments{
\item{rdd}{A (k, v) pair RDD to be sorted.}

\item{ascending}{A flag to indicate whether the sorting is ascending or descending.}

\item{numPartitions}{Number of partitions to create.}
}
\value{
An RDD where all (k, v) pair elements are sorted.
}
\description{
Sort a (k, v) pair RDD by k.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list(list(3, 1), list(2, 2), list(1, 3)))
collect(sortByKey(rdd)) # list (list(1, 3), list(2, 2), list(3, 1))
}
}

