% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sparkR.R
\name{sparkRSQL.init}
\alias{sparkRSQL.init}
\title{Initialize a new Spark Context.}
\usage{
sparkRSQL.init(jsc)
}
\arguments{
\item{jsc}{The existing JavaSparkContext created with SparkR.init()}
}
\description{
This function creates a SparkContext from an existing JavaSparkContext and
then uses it to initialize a new SQLContext
}

