% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{sql}
\alias{sql}
\title{SQL Query}
\usage{
sql(sqlctx, sqlQuery)
}
\arguments{
\item{sqlctx}{SQLContext to use}

\item{sqlQuery}{A character vector containing the SQL query}
}
\value{
DataFrame
}
\description{
Executes a SQL query using Spark, returning the result as a DataFrame.
}

