% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RDD.R
\docType{methods}
\name{top}
\alias{top}
\alias{top,RDD,RDD-method}
\alias{top,RDD,integer-method}
\title{Returns the top N elements from an RDD.}
\usage{
top(rdd, num)

\S4method{top}{RDD,integer}(rdd, num)
}
\arguments{
\item{rdd}{An RDD.}

\item{num}{Number of elements to return.}
}
\value{
The top N elements from the RDD.
}
\description{
Returns the top N elements from an RDD.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list(10, 1, 2, 9, 3, 4, 5, 6, 7))
top(rdd, 6L) # list(10, 9, 7, 6, 5, 4)
}
}

