% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{uncacheTable}
\alias{uncacheTable}
\title{Uncache Table}
\usage{
uncacheTable(sqlctx, tableName)
}
\arguments{
\item{sqlctx}{SQLContext to use}

\item{tableName}{The name of the table being uncached}
}
\value{
DataFrame
}
\description{
Removes the specified table from the in-memory cache.
}

