% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{pipeRDD}
\alias{pipeRDD}
\alias{pipeRDD,RDD,character-method}
\title{Pipes elements to a forked externel process.}
\usage{
pipeRDD(rdd, command)

\S4method{pipeRDD}{RDD,character}(rdd, command)
}
\arguments{
\item{rdd}{The RDD whose elements are piped to the forked externel process.}

\item{command}{the command to fork an externel process.}
}
\value{
a new RDD created by piping all elements to a forked externel process.
}
\description{
The same as `pipe()' in Spark.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10)
collect(pipeRDD(rdd, "cat")
Output: c("1", "2", ..., "10")
}
}

